pyproject.toml
[project]
name = "elhaz"
version = "0.1.0"
description = "A CLI and daemon for managing temporary AWS credentials via STS using boto3-refresh-session, exposing automatically refreshed credentials to shells, SDKs, tools, and more."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MPL-2.0" }
dependencies = [
    "boto3-refresh-session==10.0.4",
    "pydantic>=2.12.5",
    "pygments>=2.19.2",
    "questionary>=2.1.1",
    "ruamel-yaml>=0.19.1",
    "typer>=0.24.1",
]
authors = [
    {name = "Mike Letts",email = "michael@61418.io"}
]
maintainers = [
  {name="61418", email="general@61418.io"},
]
keywords = [
    "aws",
    "cli",
    "credentials",
    "sts",
    "refresh",
]

[project.urls]
Issues = "https://github.com/michaelthomasletts/elhaz/issues"
Repository = "https://github.com/michaelthomasletts/elhaz"

[project.scripts]
elhaz = "elhaz.cli.__main__:main"

[dependency-groups]
dev = [
    "furo>=2025.12.19",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
    "sphinx>=8.1.3",
    "sphinx-copybutton>=0.5.2",
    "sphinxext-opengraph[social-cards]>=0.13.0",
    "tomlkit>=0.14.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["LICENSE", "NOTICE", "boto3_client_cache/**"]

[tool.hatch.build.targets.wheel]
packages = ["elhaz"]

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"docs/**" = ["ALL"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
